+++ web-admin/src/features/organizations/users/OrgUsersTable.svelte
@@ -13,7 +13,6 @@
 
   export let data: OrgUser[];
   export let currentUserEmail: string;
-  export let onRemove: (email: string) => void;
   export let onSetRole: (email: string, role: string) => void;
 
   const columns: ColumnDef<OrgUser, any>[] = [
@@ -56,7 +55,6 @@
         flexRender(OrgUsersTableActionsCell, {
           email: row.original.userEmail,
           isCurrentUser: row.original.userEmail === currentUserEmail,
-          onRemove: onRemove,
         }),
       meta: {
         widthPercent: 0,
